+++ src/assetbundles/richvariables/RichVariablesAsset.php
@@ -8,9 +8,8 @@
  * @copyright Copyright (c) 2017 nystudio107
  */
 
-namespace nystudio107\richvariables\assetbundles\RichVariables;
+namespace nystudio107\richvariables\assetbundles\richvariables;
 
-use Craft;
 use craft\web\AssetBundle;
 use craft\web\assets\cp\CpAsset;
 use craft\redactor\assets\redactor\RedactorAsset;